+++ packages/ic-response-verification/src/types/certification_result.rs
@@ -9,6 +9,7 @@ const CERTIFICATION_RESULT_TYPE: &'static str = r#"
 interface CertificationResult {
   passed: boolean;
   response: CertifiedResponse;
+  verificationVersion: number;
 }
 "#;
 
@@ -21,12 +22,14 @@ pub struct CertificationResult {
     /// certification and passed verification. If verification failed then this object will be
     /// empty.
     pub response: Option<CertifiedResponse>,
+    /// The version of verification that was used to verify the response
+    pub verification_version: u16,
 }
 
 #[cfg(all(target_arch = "wasm32", feature = "js"))]
 impl From<CertificationResult> for JsValue {
     fn from(certification_result: CertificationResult) -> Self {
-        use js_sys::{Array, Boolean, Object};
+        use js_sys::{Array, Boolean, Number, Object};
 
         let passed = Boolean::from(certification_result.passed);
         let response = JsValue::from(certification_result.response);
@@ -34,7 +37,16 @@ impl From<CertificationResult> for JsValue {
         let passed_entry = Array::of2(&JsValue::from("passed"), &passed.into());
         let response_entry = Array::of2(&JsValue::from("response"), &response.into());
 
-        let result = Object::from_entries(&Array::of2(&passed_entry, &response_entry)).unwrap();
+        let verification_version = Number::from(certification_result.verification_version);
+        let verification_version_entry =
+            Array::of2(&JsValue::from("verificationVersion"), &verification_version);
+
+        let result = Object::from_entries(&Array::of3(
+            &passed_entry,
+            &response_entry,
+            &verification_version_entry,
+        ))
+        .unwrap();
 
         JsValue::from(result)
     }
@@ -49,12 +61,13 @@ mod tests {
 
     #[wasm_bindgen_test]
     fn serialize_certification_result_with_no_response() {
-        let expected = r#"{"passed":true}"#;
+        let expected = r#"{"passed":true,"verificationVersion":1}"#;
 
         assert_eq!(
             JSON::stringify(&JsValue::from(CertificationResult {
                 passed: true,
-                response: None
+                response: None,
+                verification_version: 1,
             }))
             .unwrap(),
             expected
@@ -63,7 +76,7 @@ mod tests {
 
     #[wasm_bindgen_test]
     fn serialize_certification_result_with_response() {
-        let expected = r#"{"passed":true,"response":{"statusCode":200,"body":{"0":0,"1":1,"2":2},"headers":[]}}"#;
+        let expected = r#"{"passed":true,"response":{"statusCode":200,"body":{"0":0,"1":1,"2":2},"headers":[]},"verificationVersion":2}"#;
 
         assert_eq!(
             JSON::stringify(&JsValue::from(CertificationResult {
@@ -72,7 +85,8 @@ mod tests {
                     status_code: Some(200),
                     body: vec![0, 1, 2],
                     headers: vec![],
-                })
+                }),
+                verification_version: 2,
             }))
             .unwrap(),
             expected