+++ packages/ic-response-verification/tests/v2_response_verification_happy_path.rs
@@ -71,6 +71,7 @@ mod tests {
 
         assert!(result.passed);
         assert_eq!(result.response, None);
+        assert_eq!(result.verification_version, 2);
     }
 
     #[test]
@@ -150,6 +151,7 @@ mod tests {
 
         assert!(result.passed);
         assert_eq!(result.response, Some(expected_response));
+        assert_eq!(result.verification_version, 2);
     }
 
     #[test]
@@ -237,6 +239,7 @@ mod tests {
 
         assert!(result.passed);
         assert_eq!(result.response, Some(expected_response));
+        assert_eq!(result.verification_version, 2);
     }
 
     #[test]
@@ -322,5 +325,6 @@ mod tests {
 
         assert!(result.passed);
         assert_eq!(result.response, Some(expected_response));
+        assert_eq!(result.verification_version, 2);
     }
 }