+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Feat
+
+- add verification version to certification result
+
 ## 0.1.1 (2023-04-04)
 
 ### Fix
@@ -92,5 +98,3 @@
 
 - improve wasm binary size
 - add binaryen wasm opt support
-
-## 0.0.0 (2022-09-29)